requirements.txt
stellar-sdk==8.1.1
python-dotenv==0.21.0